pull_request: allow server to choose target ref dynamically
A recent feature [1] enables the Azure Repos server to choose a dynamic
target branch when creating a PR. The REST API still requires a
targetRef parameter, but the new ignoreTargetRefAndChooseDynamically
boolean can tell the server to override this value.

When the 'az repos pr create' command does not have a target ref
specified by the user, then indicate that the default branch was used as
a "guess". We can then set ignoreTargetRefAndChooseDynamically to allow
the server to pick a different target branch.

This was tested locally with a repo that has this configuration
enabled.

[1] https://learn.microsoft.com/en-us/azure/devops/repos/git/pull-request-targets?view=azure-devops